[Server] boost 설치 전 커밋

+++ Server/NetWork/Session/TcpSession.cpp
@@ -92,7 +92,7 @@ void TcpSession::ProcessRecvForIOCP( DWORD bytesTrans )
 	PacketHeader packetHeader;
 	memcpy( &packetHeader, _recvTempBuf, sizeof( PacketHeader ) );
 
-	if ( _recvOffset < packetHeader.size )
+	if ( (int)_recvOffset < packetHeader.size )
 	{
 		_PostRecv( bytesTrans );
